feat(Participant): 딜러/플레이어가 참가자를 상속
