+++ README.md
@@ -149,6 +149,6 @@ jason: -20000
 - [x] 배팅 결과 출력
 - [x] 배팅 정보 구현
 - [x] 게임 결과 구현
-- [ ] 참가자 구현, 딜러/플레이어가 참가자를 상속
+- [x] 참가자 구현, 딜러/플레이어가 참가자를 상속
 - [ ] 베팅 금액, 수익금이 돈을 상속
 - [ ] players 일급컬렉션 구현
