+++ src/main/java/domain/Player.java
@@ -1,41 +1,16 @@
 package domain;
 
-import domain.card.Card;
+import domain.participant.Participant;
 
-public class Player {
+public class Player extends Participant {
+
     private final String name;
-    private final Hand hand;
 
     public Player(String name) {
         this.name = name;
-        hand = new Hand();
-    }
-
-    public void draw(Card card) {
-        hand.addCard(card);
-    }
-
-    public boolean isBust() {
-        return hand.isBust();
-    }
-
-    public boolean isBlackjack() {
-        return hand.isBlackjack();
-    }
-
-    public boolean isBlackjackAtFirst() {
-        return isBlackjack() && hand.size() == 2;
     }
 
     public String getName() {
         return name;
-    }
-
-    public Hand getHand() {
-        return hand;
-    }
-
-    public int getScore() {
-        return hand.getSum();
     }
 }
